UX-437 Convert navigation stories & add tests
