+++ cypress/integration/Tab.spec.js
@@ -1,7 +1,7 @@
 describe('The Tabs component', () => {
   describe('when not overflowing', () => {
     beforeEach(() => {
-      cy.visit('/iframe.html?path=/story/navigation-tabs--example-tabs');
+      cy.visit('/iframe.html?path=Tabs__basic-usage,-manual-keyboard-activation&source=false');
       // Setting viewport dimensions to avoid side effects
       cy.viewport(1400, 600);
     });
@@ -18,7 +18,7 @@ describe('The Tabs component', () => {
 
   describe('keyboard navigation', () => {
     beforeEach(() => {
-      cy.visit('/iframe.html?path=/story/navigation-tabs--example-tabs');
+      cy.visit('/iframe.html?path=Tabs__basic-usage,-manual-keyboard-activation&source=false');
       // Setting viewport dimensions to avoid side effects
       cy.viewport(1400, 600);
       cy.get('button')
@@ -68,14 +68,14 @@ describe('The Tabs component', () => {
 
   // Tests are flakey because they require layout changes
   // Remove if these become a problem
-  describe.skip('when overflowing', () => {
+  describe('when overflowing', () => {
     beforeEach(() => {
-      cy.visit('/iframe.html?path=/story/navigation-tabs--example-tabs');
-      cy.viewport(800, 600);
+      cy.visit('/iframe.html?path=Tabs__basic-usage,-manual-keyboard-activation&source=false');
+      cy.viewport(400, 600);
       cy.wait(500);
     });
 
-    it('should handle actionlist click', () => {
+    it.skip('should handle actionlist click', () => {
       cy.get('[data-id="tab-options-button"]').click();
       cy.get('[data-id="popover-content"]').should('be.visible');
       cy.get('[data-id="popover-content"] button')
